requirements.txt
pytest==8.2.2
requests==2.32.3
selene==2.0.0rc9
curlify==2.2.1
allure-pytest==2.13.5